# By Dejan Brkic (10) and others
# Via Russ Danner (16) and Dejan Brkic (5)
* 'master' of https://github.com/craftercms/studio2: (21 commits)
  Remove commented code
  CRAFTERCMS-1561 fixed issue
  CRAFTERCMS-1512 fixed issue
  CRAFTERCMS-1756 fixed issue
  CRAFTERCMS-1736 fixed issue
  CRAFTERCMS-1733 fixed issue
  CRAFTERCMS-1759 fixed issue
  CRAFTERCMS-1760: added a site component datasource
  CRAFTERCMS-1738 fixed issue
  CRAFTERCMS-1753 fixed issue
  CRAFTERCMS-1735 fixed issue
  Don't show sites until a site specific
  fix spelling error
  replace 2014 and 2015 in copyright with 2016 in groovy
  replace 2014 and 2015 in copyright with 2016
  replace 2013 in copyright with 2016
  http://issues.craftercms.org/browse/CRAFTERCMS-1723 remove duplicate message
  clean up search result default result template
  new tiny mce
  unit testing page
  ...

+++ src/main/java/org/craftercms/studio/impl/v1/service/deployment/PublishingManagerImpl.java
@@ -1,6 +1,6 @@
 /*
  * Crafter Studio Web-content authoring solution
- * Copyright (C) 2007-2015 Crafter Software Corporation.
+ * Copyright (C) 2007-2016 Crafter Software Corporation.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -407,11 +407,17 @@ public class PublishingManagerImpl implements PublishingManager {
                     if (item.getOldPath().endsWith("/" + DmConstants.INDEX_FILE)) {
                         boolean hasRenamedChildren = false;
                         String fullPath = contentService.expandRelativeSitePath(item.getSite(), item.getOldPath().replace("/" + DmConstants.INDEX_FILE, ""));
-                        RepositoryItem[] children = contentRepository.getContentChildren(fullPath);
-                        if (children.length < 2) {
-                            deployer.deleteFile(item.getSite(), item.getOldPath().replace("/" + DmConstants.INDEX_FILE, ""));
-                        } else {
-                            hasRenamedChildren = true;
+                        if (contentService.contentExists(fullPath)) {
+                            try {
+                                RepositoryItem[] children = contentRepository.getContentChildren(fullPath);
+                                if (children.length < 2) {
+                                    deployer.deleteFile(item.getSite(), item.getOldPath().replace("/" + DmConstants.INDEX_FILE, ""));
+                                } else {
+                                    hasRenamedChildren = true;
+                                }
+                            } catch (Exception exc) {
+                                LOGGER.info("Error while checking children for moved content site " + item.getSite() + " old path " + item.getOldPath());
+                            }
                         }
                         if (!hasRenamedChildren) {
                             contentService.deleteContent(item.getSite(), item.getOldPath().replace("/" + DmConstants.INDEX_FILE, ""), false, item.getUser());
